Made tests that could be independent, more independent
