+++ test/standalone/test-options-credentials.js
@@ -20,7 +20,6 @@ var assert = require('assert');
 var nock = require('nock');
 var nocks = require('../nocks.js');
 var extend = require('extend');
-var defaultOptions = {};
 var config = require('../../src/agent/config.js');
 var Debuglet = require('../../src/agent/debuglet.js');
 
@@ -44,34 +43,33 @@ describe('test-options-credentials', function() {
   });
 
   it('should use options.projectId in preference to the environment variable',
-     function(done) {
-       process.env.GCLOUD_PROJECT = 'should-not-be-used';
+    function(done) {
+      process.env.GCLOUD_PROJECT = 'should-not-be-used';
 
-       var options = extend({}, defaultOptions, {
-         projectId: 'project-via-options',
-         credentials: require('../fixtures/gcloud-credentials.json')
-       });
-       var debug = require('../..')(options);
+      var options = extend({}, {
+        projectId: 'project-via-options',
+        credentials: require('../fixtures/gcloud-credentials.json')
+      });
+      var debug = require('../..')(options);
 
-       // TODO: also make sure we don't request the project from metadata
-       // service.
+      // TODO: also make sure we don't request the project from metadata
+      // service.
+      var scope = nocks.oauth2();
+      nocks.register(function(body) {
+        assert.ok(body.debuggee);
+        assert.equal(body.debuggee.project, 'project-via-options');
+        scope.done();
+        setImmediate(done);
+        return true;
+      });
 
-        var scope = nocks.oauth2();
-        nocks.register(function(body) {
-          assert.ok(body.debuggee);
-          assert.equal(body.debuggee.project, 'project-via-options');
-          scope.done();
-          setImmediate(done);
-          return true;
-        });
-
-        debuglet = new Debuglet(debug, config);
-       debuglet.start();
-     });
+      debuglet = new Debuglet(debug, config);
+      debuglet.start();
+    });
 
   it('should use the keyFilename field of the options object', function(done) {
     var credentials = require('../fixtures/gcloud-credentials.json');
-    var options = extend({}, defaultOptions, {
+    var options = extend({}, {
       projectId: 'fake-project',
       keyFilename: path.join('test', 'fixtures', 'gcloud-credentials.json')
     });
@@ -93,7 +91,7 @@ describe('test-options-credentials', function() {
   });
 
   it('should use the credentials field of the options object', function(done) {
-    var options = extend({}, defaultOptions, {
+    var options = extend({}, {
       projectId: 'fake-project',
       credentials: require('../fixtures/gcloud-credentials.json')
     });
@@ -122,7 +120,7 @@ describe('test-options-credentials', function() {
       refresh_token: 'c',
       type: 'authorized_user'
     };
-    var options = extend({}, defaultOptions, {
+    var options = extend({}, {
       projectId: 'fake-project',
       keyFilename: path.join('test', 'fixtures', 'gcloud-credentials.json'),
       credentials: credentials
